rory_personal_assistant/streamlit_requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
networkx>=3.1

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Vector database (with fallback handling)
chromadb>=0.4.15

# Optional: PDF processing (fallback available)
PyMuPDF>=1.23.0

# Optional: Sentence transformers (fallback available)
sentence-transformers>=2.2.2

# OCR libraries for enhanced PDF processing (optional)
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0
